// libs/features/establishment/src/lib/shared/enums/navigation-indicator-enum.ts
export enum NavigationIndicatorEnum {
  COMPLETE_PROACTIVE_SAVE = 1,
  COMPLETE_PROACTIVE_PREVIOUS_SAVE = 2,
  COMPLETE_PROACTIVE_SUBMIT = 3,
  CSR_SAVE = 4,
  CSR_SUBMIT = 5,
  FIRST_VALIDATOR_SAVE = 6,
  FIRST_VALIDATOR_SUBMIT = 7,
  COMPLETE_PROACTIVE_RESUBMIT = 8,
  CSR_CHANGE_BASIC_DETAILS_SUBMIT = 9,
  ESTADMIN_CHANGE_BASIC_DETAILS_SUBMIT = 10,
  VALIDATOR_CHANGE_BASIC_DETAILS_SUBMIT = 11,
  CSR_CHANGE_IDENTIFIER_DETAILS_SUBMIT = 12,
  ESTADMIN_CHANGE_IDENTIFIER_DETAILS_SUBMIT = 13,
  VALIDATOR_CHANGE_IDENTIFIER_DETAILS_SUBMIT = 14,
  CSR_CHANGE_BANK_DETAILS_SUBMIT = 15,
  ESTADMIN_CHANGE_ACCOUNT_DETAILS_SUBMIT = 16,
  VALIDATOR_CHANGE_BANK_DETAILS_SUBMIT = 17,
  CSR_CHANGE_OWNER_SAVE = 18,
  ESTADMIN_CHANGE_OWNER_SAVE = 19,
  VALIDATOR_CHANGE_OWNER_SAVE = 20,
  CSR_CHANGE_OWNER_FINAL_SUBMIT = 21,
  EST_ADMIN_CHANGE_OWNER_FINAL_SUBMIT = 22,
  ADMIN_REENTER_CHANGE_OWNER_FINAL_SUBMIT = 23,
  VALIDATOR_CHANGE_OWNER_FINAL_SUBMIT = 24,
  CSR_CHANGE_CONTACT_DETAILS_SUBMIT = 25,
  VALIDATOR_CHANGE_CONTACT_DETAILS_SUBMIT = 26,
  ESTADMIN_CHANGE_CONTACT_DETAILS_SUBMIT = 27,
  CSR_CHANGE_ADDRESS_DETAILS_SUBMIT = 28,
  ESTADMIN_CHANGE_ADDRESS_DETAILS_SUBMIT = 29,
  VALIDATOR_CHANGE_ADDRESS_DETAILS_SUBMIT = 30,
  CSR_LEGAL_ENTITY_CHANGE_SUBMIT = 31,
  ADD_OWNER_LEGAL_ENTITY_CHANGE = 32,
  VALIDATOR_LEGAL_ENTITY_CHANGE_SUBMIT = 33,
  CSR_LEGAL_ENTITY_CHANGE_FINAL_SUBMIT = 34,
  VALIDATOR_LEGAL_ENTITY_CHANGE_FINAL_SUBMIT = 35,
  CSR_CHANGE_BRANCH_TO_MAIN_SUBMIT = 36,
  ESTADMIN_CHANGE_BRANCH_TO_MAIN_SUBMIT = 37,
  VALIDATOR_CHANGE_BRANCH_TO_MAIN_SUBMIT = 38,
  CSR_CHANGE_BRANCH_TO_MAIN_FINAL_SUBMIT = 39,
  ADMIN_CHANGE_BRANCH_TO_MAIN_FINAL_SUBMIT = 40,
  VALIDATOR_CHANGE_BRANCH_TO_MAIN_FINAL_SUBMIT = 41,
  CSR_ESTABLISHMENT_TRANSFER_LINK_FINAL_SUBMIT = 42,
  VALIDATOR_ESTABLISHMENT_TRANSFER_LINK_FINAL_SUBMIT = 43,
  CSR_ESTABLISHMENT_TRANSFER_DELINK_FINAL_SUBMIT = 44,
  VALIDATOR_ESTABLISHMENT_TRANSFER_DELINK_FINAL_SUBMIT = 45,
  CSR_ESTABLISHMENT_TRANSFER_LINK_SUBMIT = 46,
  VALIDATOR_ESTABLISHMENT_TRANSFER_LINK_SUBMIT = 47,
  CSR_ESTABLISHMENT_TRANSFER_DELINK_SUBMIT = 48,
  VALIDATOR_ESTABLISHMENT_TRANSFER_DELINK_SUBMIT = 49,
  CSR_REPLACE_SUPER_ADMIN = 50,
  VALIDATOR_REPLACE_SUPER_ADMIN = 51,
  CSR_REPLACE_SUPER_ADMIN_FINAL_SUBMIT = 52,
  CSR_ADD_SUPER_ADMIN_SUBMIT = 74,
  CSR_ADD_SUPER_ADMIN_FINAL_SUBMIT = 76,
  VALIDATOR_ADD_SUPER_ADMIN_SUBMIT = 75,
  VALIDATOR_ADD_SUPER_ADMIN_FINAL_SUBMIT = 77,
  VALIDATOR_REPLACE_SUPER_ADMIN_FINAL_SUBMIT = 53,
  ESTADMIN_REPLACE_SUPER_ADMIN = 54,
  ESTADMIN_REPLACE_SUPER_ADMIN_FINAL_SUBMIT = 55,
  CSR_TERMINATE_ESTABLISHMENT_SUBMIT = 56,
  VALIDATOR_TERMINATE_ESTABLISHMENT_SUBMIT = 57,
  GOL_TERMINATE_ESTABLISHMENT_SUBMIT = 59,
  GOL_VALIDATOR_TERMINATE_EST_SUBMIT = 61,
  ESTADMIN_REENTER_BANK = 60,
  FLAG_EST_SUBMIT = 62,
  FLAG_EST_FINAL_SUBMIT = 63,
  CUSTOMER_CARE_MODIFY_FLAG = 64,
  CUSTOMER_CARE_MODIFY_FLAG_SUBMIT = 65,
  CUSTOMER_CARE_FLAG_EST_FINAL_SUBMIT = 66,
  CUSTOMER_CARE_REENTER_MODIFY_FLAG_SUBMIT = 67,
  ADMIN_LEGAL_ENTITY_CHANGE_SUBMIT = 68,
  ADMIN_LEGAL_ENTITY_CHANGE_FINAL_SUBMIT = 69,
  ADMIN_REENTER_LEGAL_ENTITY_FINAL_SUBMIT = 70,
  CSR_CHANGE_LATE_FEE_DETAILS_SUBMIT = 72,
  VALIDATOR_CHANGE_LATE_FEE_DETAILS_SUBMIT = 73,
  CSR_MOF_PAYMENT_SUBMIT = 78,
  CSR_MOF_PAYMENT_FINAL_SUBMIT = 79,
  VALIDATOR_MOF_PAYMENT_SUBMIT = 80,
  VALIDATOR_MOF_PAYMENT_FINAL_SUBMIT = 81,
  ESTADMIN_ESTABLISHMENT_TRANSFER_DELINK_SUBMIT = 82,
  ESTADMIN_ESTABLISHMENT_TRANSFER_DELINK_FINAL_SUBMIT = 83,
  ESTADMIN_ESTABLISHMENT_TRANSFER_LINK_SUBMIT = 84,
  ESTADMIN_ESTABLISHMENT_TRANSFER_LINK_FINAL_SUBMIT = 85,
  ADMIN_LEGAL_ENTITY_DIRECT_CHANGE_SUBMIT = 86,
  ESTADMIN_REPLACE_ADMIN_DIRECT_CHANGE_SUBMIT = 87,
  ESTADMIN_CHANGE_OWNER_DIRECT_CHANGE = 88,
  ESTADMIN_ADD_BANK_DETAILS_DIRECT_CHANGE = 95,
  REOPEN_EST_FIRST_SAVE_AND_NEXT = 96,
  REOPEN_EST_SECOND_SAVE_AND_NEXT = 97,
  REOPEN_FINAL_SUBMIT = 98,
  VALIDATOR_REOPEN_ESTABLISHMENT_FINAL_SUBMIT = 99,
  VALIDATOR_REOPEN_EST_FIRST_SAVE_AND_NEXT = 101,
  VALIDATOR_REOPEN_EST_SECOND_SAVE_AND_NEXT = 102,
  COMPLETE_PROACTIVE_DIRECT_CHANGE_SUBMIT = 103,
  ESTADMIN_CHANGE_OWNER_REENTER_SAVE = 104
}